WIP: Add Scene builder and refactor PlatformFactory
