+++ src/platform/lumin-runtime/elements/builders/prism-builder.js
@@ -15,41 +15,51 @@ import { logError } from '../../utilities/logger.js';
 
 export class PrismBuilder extends ElementBuilder {
   constructor () {
-      super();
+    super();
 
-      this._propertyDescriptors['position'] = new ArrayProperty('localPosition', 'setLocalPosition', true, 'vec3');
-      this._propertyDescriptors['excludeFromAutoFocus'] = new PrimitiveTypeProperty('excludeFromAutoFocus', 'setExcludeFromAutoFocus', true, 'boolean');
-      this._propertyDescriptors['handGestureFilterConfidenceLevel'] = new PrimitiveTypeProperty('handGestureFilterConfidenceLevel', 'setHandGestureFilterConfidenceLevel', true, 'number');
-      this._propertyDescriptors['setHandGestureFilterPollRate'] = new PrimitiveTypeProperty('handGestureFilterPollRate', 'setHandGestureFilterPollRate', true, 'number');
-      this._propertyDescriptors['setHandGestureFilterPositionDelta'] = new PrimitiveTypeProperty('handGestureFilterPositionDelta', 'setHandGestureFilterPositionDelta', true, 'number');
-      this._propertyDescriptors['setHandGestureHoverDistance'] = new PrimitiveTypeProperty('handGestureHoverDistance', 'setHandGestureHoverDistance', true, 'number');
-      this._propertyDescriptors['setHandGestureTouchDistance'] = new PrimitiveTypeProperty('handGestureTouchDistance', 'setHandGestureTouchDistance', true, 'number');
-      this._propertyDescriptors['setPhysicsEnabled'] = new PrimitiveTypeProperty('physicsEnabled','setPhysicsEnabled', true, 'boolean');
-      this._propertyDescriptors['setPhysicsPaused'] = new PrimitiveTypeProperty('physicsPaused', 'setPhysicsPaused', true, 'boolean');
-      this._propertyDescriptors['setPhysicsWorldMeshEnabled'] = new PrimitiveTypeProperty('physicsWorldMeshEnabled', 'setPhysicsWorldMeshEnabled', true, 'boolean');
-      this._propertyDescriptors['setPrismBloomStrength'] = new PrimitiveTypeProperty('prismBloomStrength', 'setPrismBloomStrength', true, 'number');
-      this._propertyDescriptors['setVolumeBloomStrength'] = new PrimitiveTypeProperty('volumeBloomStrength', 'setVolumeBloomStrength', true, 'number');
-      this._propertyDescriptors['trackHandGesture'] = new PrimitiveTypeProperty('trackHandGesture', 'trackHandGesture', false, 'number');
-      this._propertyDescriptors['trackingAutoHapticOnGesture'] = new PrimitiveTypeProperty('trackingAutoHapticOnGesture', 'trackingAutoHapticOnGesture', false, 'number');
+    this._propertyDescriptors['position'] = new ArrayProperty('localPosition', 'setLocalPosition', true, 'vec3');
+    this._propertyDescriptors['excludeFromAutoFocus'] = new PrimitiveTypeProperty('excludeFromAutoFocus', 'setExcludeFromAutoFocus', true, 'boolean');
+    this._propertyDescriptors['handGestureFilterConfidenceLevel'] = new PrimitiveTypeProperty('handGestureFilterConfidenceLevel', 'setHandGestureFilterConfidenceLevel', true, 'number');
+    this._propertyDescriptors['setHandGestureFilterPollRate'] = new PrimitiveTypeProperty('handGestureFilterPollRate', 'setHandGestureFilterPollRate', true, 'number');
+    this._propertyDescriptors['setHandGestureFilterPositionDelta'] = new PrimitiveTypeProperty('handGestureFilterPositionDelta', 'setHandGestureFilterPositionDelta', true, 'number');
+    this._propertyDescriptors['setHandGestureHoverDistance'] = new PrimitiveTypeProperty('handGestureHoverDistance', 'setHandGestureHoverDistance', true, 'number');
+    this._propertyDescriptors['setHandGestureTouchDistance'] = new PrimitiveTypeProperty('handGestureTouchDistance', 'setHandGestureTouchDistance', true, 'number');
+    this._propertyDescriptors['setPhysicsEnabled'] = new PrimitiveTypeProperty('physicsEnabled','setPhysicsEnabled', true, 'boolean');
+    this._propertyDescriptors['setPhysicsPaused'] = new PrimitiveTypeProperty('physicsPaused', 'setPhysicsPaused', true, 'boolean');
+    this._propertyDescriptors['setPhysicsWorldMeshEnabled'] = new PrimitiveTypeProperty('physicsWorldMeshEnabled', 'setPhysicsWorldMeshEnabled', true, 'boolean');
+    this._propertyDescriptors['setPrismBloomStrength'] = new PrimitiveTypeProperty('prismBloomStrength', 'setPrismBloomStrength', true, 'number');
+    this._propertyDescriptors['setVolumeBloomStrength'] = new PrimitiveTypeProperty('volumeBloomStrength', 'setVolumeBloomStrength', true, 'number');
+    this._propertyDescriptors['trackHandGesture'] = new PrimitiveTypeProperty('trackHandGesture', 'trackHandGesture', false, 'number');
+    this._propertyDescriptors['trackingAutoHapticOnGesture'] = new PrimitiveTypeProperty('trackingAutoHapticOnGesture', 'trackingAutoHapticOnGesture', false, 'number');
 
     this._propertyDescriptors['onDestroy'] = new EventProperty('onDestroy', 'onDestroySub', true);
   }
 
-  create(prism, properties) {
-    logError('Prism should be created by the App, no by PrismBuilder!');
+  create(app, properties) {
+    const prism = app.addPrism(properties);
+
+    Object.defineProperty(prism, 'addChild$Universal', {
+      enumerable: true,
+      writable: false,
+      configurable: false,
+      value: (child) => console.log('Prism is not supposed to add children !')
+    });
+
+    return prism;
   }
 
   validate(prism, oldProperties, newProperties) {
-
+    console.log('PrismBuilder.validate has not been implemented yet');
   }
 
   update(prism, oldProperties, newProperties) {
-
+    console.log('PrismBuilder.validate has not been implemented yet');
   }
 
   trackHandGesture(prism, oldProperties, newProperties) {
     // startTrackHandGesture
     // stopTrackHandGesture
+    console.log('PrismBuilder.validate has not been implemented yet');
   }
 
   trackingAutoHapticOnGesture(prism, oldProperties, newProperties) {
